fastboot: improved fix for writing to sd cards
There's a bit to check for SD vs MMC (and macros for it:  IS_MMC() and IS_SD()).
The IS_MMC() macro checks bit 30 which is set for eMMC, and IS_SD() checks bit 31
which is for SD cards. This change checks for eMMC vs SD, and also if there's a boot
partition available if it's emmc (which should handle the case where maybe an emmc
part doesn't have a boot partition).

This fixes "fastboot: support for block write and boot partition access".

Signed-off-by: Pat Kusbel <pat@kusbel.com>
Signed-off-by: Troy Kisky <troy.kisky@boundarydevices.com>